wip(#1099): extract lower_call/options + per-API sub-split (notification/abort/fetch)
