+++ crates/perry-codegen/src/lower_call/mod.rs
@@ -42,11 +42,22 @@ mod method_override;
 mod native;
 mod native_table;
 mod new;
+mod options;
 mod ui_styling;
 use buffer_intrinsic::try_emit_buffer_read_intrinsic;
 use builtin::lower_builtin_new;
 use jsx::try_rewrite_perry_tui_jsx_intrinsic;
 use method_override::emit_own_method_override_check;
+// `options/` (#1099): the options-object-literal lowering family,
+// split by native-API surface (notification / abort / fetch) under
+// `options/`. Bring the per-surface entry points + shared helpers
+// into this module's scope so the existing `super::<name>` call
+// sites in sibling submodules (builtin/native/ui_styling) keep
+// resolving unchanged after the split.
+use options::{
+    build_headers_from_object, get_raw_string_ptr, lower_abort_controller_call,
+    lower_fetch_native_method, lower_notification_schedule,
+};
 // `native_table.rs` (#1099): the ~5k-row `NATIVE_MODULE_TABLE` data +
 // arg/ret kind types. The dispatch consumers below
 // (`native_module_lookup`, `lower_native_module_dispatch`) stay here
@@ -61,6 +72,9 @@ pub(crate) use native::lower_native_method_call;
 // (codegen.rs / expr.rs / stmt.rs) so `crate::lower_call::lower_new`
 // etc. keep resolving after the split.
 pub(crate) use new::{apply_field_initializers_recursive, lower_new, FieldInitMode};
+// `extract_options_fields` is consumed by `expr.rs` as
+// `crate::lower_call::extract_options_fields` — keep that path stable.
+pub(crate) use options::extract_options_fields;
 // `iter_native_module_table` is consumed by `lib.rs`'s public manifest
 // API as `lower_call::iter_native_module_table` — keep that path stable.
 pub(crate) use native_table::iter_native_module_table;
@@ -71,7 +85,7 @@ use crate::type_analysis::{
     is_array_expr, is_global_constructor_expr, is_map_expr, is_promise_expr, is_set_expr,
     is_string_expr, receiver_class_name,
 };
-use crate::types::{DOUBLE, I32, I64, I8, PTR, VOID};
+use crate::types::{DOUBLE, I32, I64, I8, PTR};
 
 /// Lower a `Call` expression. Two shapes are supported:
 /// 1. `FuncRef(id)(args...)` — direct call to a user function by HIR id.
@@ -3170,15 +3184,6 @@ pub(crate) fn lower_call(ctx: &mut FnCtx<'_>, callee: &Expr, args: &[Expr]) -> R
 /// For both shapes we chain a get + push + write-back so reallocations
 /// are reflected in the source storage.
 
-/// Extract a raw string pointer (i64) from a NaN-boxed JSValue via the
-/// unified helper. Handles string literals, concat results, and any
-/// other expression that produces a NaN-boxed double.
-pub(super) fn get_raw_string_ptr(ctx: &mut FnCtx<'_>, e: &Expr) -> Result<String> {
-    let v = lower_expr(ctx, e)?;
-    let blk = ctx.block();
-    Ok(blk.call(I64, "js_get_string_pointer_unified", &[(DOUBLE, &v)]))
-}
-
 /// Issue #185 Phase C step 2: apply an inline `style: { ... }` object
 /// to a freshly-created widget handle by destructuring the object
 /// literal at HIR time and emitting a sequence of setter calls.
@@ -3197,1121 +3202,9 @@ pub(super) fn get_raw_string_ptr(ctx: &mut FnCtx<'_>, e: &Expr) -> Result<String
 /// Mirrors the App({...}) destructure pattern in this file:
 /// `extract_options_fields` returns the props, then per-key routing.
 
-/// Build a Headers handle from an inline object literal `{ "k": "v", ... }`.
-/// Returns the f64 handle (raw numeric, not NaN-boxed).
-pub(super) fn build_headers_from_object(
-    ctx: &mut FnCtx<'_>,
-    props: &[(String, Expr)],
-) -> Result<String> {
-    let h = ctx.block().call(DOUBLE, "js_headers_new", &[]);
-    for (k, vexpr) in props {
-        let key_expr = Expr::String(k.clone());
-        let key_ptr = get_raw_string_ptr(ctx, &key_expr)?;
-        let val_ptr = get_raw_string_ptr(ctx, vexpr)?;
-        ctx.block().call(
-            DOUBLE,
-            "js_headers_set",
-            &[(DOUBLE, &h), (I64, &key_ptr), (I64, &val_ptr)],
-        );
-    }
-    Ok(h)
-}
-
-/// Phase 3 compat: extract `{key: value, ...}` pairs from an options
-/// argument in a form that works whether the options literal reached us
-/// as a plain `Expr::Object(props)` (pre-Phase-3 / spread/dynamic shapes)
-/// or as an `Expr::New { class_name: "__AnonShape_N", args }` (Phase 3's
-/// closed-shape synthesis path). For the anon-class form, `ctx.classes`
-/// carries the class with its synthesized constructor — we pair each
-/// constructor param name with its positional arg to recover the literal's
-/// (key, value) view.
-///
-/// Returns `None` when the expression is neither shape — callers should
-/// fall through to whatever they did before when the 2nd arg wasn't an
-/// inline object.
-pub(crate) fn extract_options_fields(ctx: &FnCtx<'_>, e: &Expr) -> Option<Vec<(String, Expr)>> {
-    match e {
-        Expr::Object(props) => Some(props.clone()),
-        Expr::New {
-            class_name, args, ..
-        } if class_name.starts_with("__AnonShape_") => {
-            let class = ctx.classes.get(class_name)?;
-            let ctor = class.constructor.as_ref()?;
-            if ctor.params.len() != args.len() {
-                return None;
-            }
-            let pairs: Vec<(String, Expr)> = ctor
-                .params
-                .iter()
-                .zip(args.iter())
-                .map(|(param, arg)| (param.name.clone(), arg.clone()))
-                .collect();
-            Some(pairs)
-        }
-        _ => None,
-    }
-}
-
-/// Lower `notificationSchedule({ id, title, body, trigger })` (#96). Switches
-/// on `trigger.type` (which must be a string literal at the call site so we
-/// can pick the right runtime fn at compile time) and emits a flat-arg call
-/// to one of three runtime fns:
-/// - `interval` → `perry_system_notification_schedule_interval(id, title, body, seconds, repeats)`
-/// - `calendar` → `perry_system_notification_schedule_calendar(id, title, body, timestamp_ms)`
-/// - `location` → `perry_system_notification_schedule_location(id, title, body, lat, lon, radius)`
-///
-/// `repeats` is passed as a NaN-boxed JS value; the runtime calls
-/// `js_is_truthy` to coerce. Missing fields default to 0.0.
-pub(super) fn lower_notification_schedule(ctx: &mut FnCtx<'_>, args: &[Expr]) -> Result<String> {
-    if args.len() != 1 {
-        bail!(
-            "notificationSchedule(...) takes one argument: \
-             {{ id, title, body, trigger }} (got {} args)",
-            args.len()
-        );
-    }
-    let Some(props) = extract_options_fields(ctx, &args[0]) else {
-        bail!(
-            "notificationSchedule(...) requires an inline object literal: \
-             {{ id: ..., title: ..., body: ..., trigger: {{ ... }} }}"
-        );
-    };
-
-    let mut id_ptr: Option<String> = None;
-    let mut title_ptr: Option<String> = None;
-    let mut body_ptr: Option<String> = None;
-    let mut trigger: Option<Vec<(String, Expr)>> = None;
-
-    for (key, val) in &props {
-        match key.as_str() {
-            "id" => {
-                let v = lower_expr(ctx, val)?;
-                let blk = ctx.block();
-                id_ptr = Some(unbox_to_i64(blk, &v));
-            }
-            "title" => {
-                let v = lower_expr(ctx, val)?;
-                let blk = ctx.block();
-                title_ptr = Some(unbox_to_i64(blk, &v));
-            }
-            "body" => {
-                let v = lower_expr(ctx, val)?;
-                let blk = ctx.block();
-                body_ptr = Some(unbox_to_i64(blk, &v));
-            }
-            "trigger" => {
-                let Some(tprops) = extract_options_fields(ctx, val) else {
-                    bail!(
-                        "notificationSchedule: `trigger` must be an inline object literal \
-                         like `{{ type: \"interval\", seconds: 60 }}`"
-                    );
-                };
-                trigger = Some(tprops);
-            }
-            _ => {
-                let _ = lower_expr(ctx, val)?;
-            }
-        }
-    }
-
-    let id_ptr = id_ptr
-        .ok_or_else(|| anyhow::anyhow!("notificationSchedule: missing required field `id`"))?;
-    let title_ptr = title_ptr
-        .ok_or_else(|| anyhow::anyhow!("notificationSchedule: missing required field `title`"))?;
-    let body_ptr = body_ptr
-        .ok_or_else(|| anyhow::anyhow!("notificationSchedule: missing required field `body`"))?;
-    let trigger = trigger
-        .ok_or_else(|| anyhow::anyhow!("notificationSchedule: missing required field `trigger`"))?;
-
-    let mut trigger_type: Option<String> = None;
-    for (k, v) in &trigger {
-        if k == "type" {
-            match v {
-                Expr::String(s) => trigger_type = Some(s.clone()),
-                _ => bail!(
-                    "notificationSchedule: `trigger.type` must be a string literal \
-                     (one of \"interval\", \"calendar\", \"location\") at the call site"
-                ),
-            }
-            break;
-        }
-    }
-    let trigger_type = trigger_type.ok_or_else(|| {
-        anyhow::anyhow!("notificationSchedule: missing required field `trigger.type`")
-    })?;
-
-    match trigger_type.as_str() {
-        "interval" => {
-            let mut seconds: String = "0.0".to_string();
-            let mut repeats: String = double_literal(f64::from_bits(crate::nanbox::TAG_FALSE));
-            for (k, v) in &trigger {
-                match k.as_str() {
-                    "type" => {}
-                    "seconds" => seconds = lower_expr(ctx, v)?,
-                    "repeats" => repeats = lower_expr(ctx, v)?,
-                    _ => {
-                        let _ = lower_expr(ctx, v)?;
-                    }
-                }
-            }
-            ctx.pending_declares.push((
-                "perry_system_notification_schedule_interval".to_string(),
-                VOID,
-                vec![I64, I64, I64, DOUBLE, DOUBLE],
-            ));
-            ctx.block().call_void(
-                "perry_system_notification_schedule_interval",
-                &[
-                    (I64, &id_ptr),
-                    (I64, &title_ptr),
-                    (I64, &body_ptr),
-                    (DOUBLE, &seconds),
-                    (DOUBLE, &repeats),
-                ],
-            );
-        }
-        "calendar" => {
-            let mut timestamp_ms: String = "0.0".to_string();
-            for (k, v) in &trigger {
-                match k.as_str() {
-                    "type" => {}
-                    "date" => timestamp_ms = lower_expr(ctx, v)?,
-                    _ => {
-                        let _ = lower_expr(ctx, v)?;
-                    }
-                }
-            }
-            ctx.pending_declares.push((
-                "perry_system_notification_schedule_calendar".to_string(),
-                VOID,
-                vec![I64, I64, I64, DOUBLE],
-            ));
-            ctx.block().call_void(
-                "perry_system_notification_schedule_calendar",
-                &[
-                    (I64, &id_ptr),
-                    (I64, &title_ptr),
-                    (I64, &body_ptr),
-                    (DOUBLE, &timestamp_ms),
-                ],
-            );
-        }
-        "location" => {
-            let mut lat: String = "0.0".to_string();
-            let mut lon: String = "0.0".to_string();
-            let mut radius: String = "0.0".to_string();
-            for (k, v) in &trigger {
-                match k.as_str() {
-                    "type" => {}
-                    "latitude" => lat = lower_expr(ctx, v)?,
-                    "longitude" => lon = lower_expr(ctx, v)?,
-                    "radius" => radius = lower_expr(ctx, v)?,
-                    _ => {
-                        let _ = lower_expr(ctx, v)?;
-                    }
-                }
-            }
-            ctx.pending_declares.push((
-                "perry_system_notification_schedule_location".to_string(),
-                VOID,
-                vec![I64, I64, I64, DOUBLE, DOUBLE, DOUBLE],
-            ));
-            ctx.block().call_void(
-                "perry_system_notification_schedule_location",
-                &[
-                    (I64, &id_ptr),
-                    (I64, &title_ptr),
-                    (I64, &body_ptr),
-                    (DOUBLE, &lat),
-                    (DOUBLE, &lon),
-                    (DOUBLE, &radius),
-                ],
-            );
-        }
-        other => bail!(
-            "notificationSchedule: unknown trigger.type \"{}\" \
-             (expected one of \"interval\", \"calendar\", \"location\")",
-            other
-        ),
-    }
-
-    Ok(double_literal(0.0))
-}
-
 /// Lower `new ClassName(args)` for the built-in Web classes that don't
 /// live in `ctx.classes`. Returns `Ok(None)` if the class isn't one we
 /// handle here (caller should fall through to the default path).
-
-/// Returns `true` if the expression statically resolves to an
-/// `AbortController`-typed value (either a local whose declared type
-/// is `Named("AbortController")` or a `new AbortController()` call).
-pub(super) fn is_abort_controller_expr(ctx: &FnCtx<'_>, e: &Expr) -> bool {
-    match e {
-        Expr::New { class_name, .. } => class_name == "AbortController",
-        Expr::LocalGet(id) => matches!(
-            ctx.local_types.get(id),
-            Some(HirType::Named(n)) if n == "AbortController"
-        ),
-        _ => false,
-    }
-}
-
-/// Lower AbortController / AbortSignal method calls:
-/// - `controller.abort(reason?)`
-/// - `controller.signal.addEventListener("abort", cb)`
-/// - `AbortSignal.timeout(ms)` (static)
-///
-/// Returns `None` if the call shape doesn't match one of the handled
-/// patterns — caller falls through to the generic dispatch.
-pub(super) fn lower_abort_controller_call(
-    ctx: &mut FnCtx<'_>,
-    object: &Expr,
-    property: &str,
-    args: &[Expr],
-) -> Result<Option<String>> {
-    // ── AbortSignal.timeout(ms) static ──
-    if property == "timeout" {
-        if let Expr::GlobalGet(_) = object {
-            // Can't distinguish AbortSignal.timeout from other globals
-            // without more context — skip.
-        }
-    }
-    // Static `AbortSignal.timeout(ms)` — matched via a PropertyGet on a
-    // GlobalGet-shaped object isn't quite right because GlobalGet has
-    // no name; best we can do is detect by property name "timeout" and
-    // the local-isn't-a-known-thing. Skip for now.
-
-    // ── controller.abort(reason?) ──
-    if property == "abort" && is_abort_controller_expr(ctx, object) {
-        let recv_box = lower_expr(ctx, object)?;
-        let blk = ctx.block();
-        let ctrl_handle = unbox_to_i64(blk, &recv_box);
-        if args.is_empty() {
-            blk.call_void("js_abort_controller_abort", &[(I64, &ctrl_handle)]);
-        } else {
-            let reason = lower_expr(ctx, &args[0])?;
-            let blk = ctx.block();
-            blk.call_void(
-                "js_abort_controller_abort_reason",
-                &[(I64, &ctrl_handle), (DOUBLE, &reason)],
-            );
-        }
-        return Ok(Some(double_literal(f64::from_bits(
-            crate::nanbox::TAG_UNDEFINED,
-        ))));
-    }
-
-    // ── controller.signal.addEventListener("abort", cb) ──
-    if property == "addEventListener" && args.len() >= 2 {
-        if let Expr::PropertyGet {
-            object: inner_obj,
-            property: inner_prop,
-        } = object
-        {
-            if inner_prop == "signal" && is_abort_controller_expr(ctx, inner_obj) {
-                let ctrl_box = lower_expr(ctx, inner_obj)?;
-                let blk = ctx.block();
-                let ctrl_handle = unbox_to_i64(blk, &ctrl_box);
-                // Get the signal pointer.
-                let signal_handle =
-                    blk.call(I64, "js_abort_controller_signal", &[(I64, &ctrl_handle)]);
-                let evt = lower_expr(ctx, &args[0])?;
-                let listener = lower_expr(ctx, &args[1])?;
-                let blk = ctx.block();
-                blk.call_void(
-                    "js_abort_signal_add_listener",
-                    &[(I64, &signal_handle), (DOUBLE, &evt), (DOUBLE, &listener)],
-                );
-                return Ok(Some(double_literal(f64::from_bits(
-                    crate::nanbox::TAG_UNDEFINED,
-                ))));
-            }
-        }
-    }
-
-    Ok(None)
-}
-
-/// Dispatch for the Web Fetch API family: Response/Headers/Request
-/// methods and property getters. Called before the generic
-/// `lower_native_method_call` path so static factories
-/// (`Response.json(v)`) also land here. Returns `Ok(None)` if the
-/// (module, method) combination isn't handled.
-///
-/// Handle ABI note: Response/Headers/Request handles are plain numeric
-/// doubles (ids into the runtime's registry), not NaN-boxed pointers.
-/// Most runtime functions take the handle as f64; status/statusText/
-/// ok/text/json take i64 and we convert via `fptosi`.
-pub(super) fn lower_fetch_native_method(
-    ctx: &mut FnCtx<'_>,
-    module: &str,
-    method: &str,
-    object: Option<&Expr>,
-    args: &[Expr],
-) -> Result<Option<String>> {
-    // ── Response static factories (no receiver) ──
-    if module == "fetch" && object.is_none() {
-        match method {
-            "static_json" => {
-                let v = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                let handle = ctx
-                    .block()
-                    .call(DOUBLE, "js_response_static_json", &[(DOUBLE, &v)]);
-                return Ok(Some(handle));
-            }
-            "static_redirect" => {
-                let url_ptr = if !args.is_empty() {
-                    get_raw_string_ptr(ctx, &args[0])?
-                } else {
-                    "0".to_string()
-                };
-                let status = if args.len() >= 2 {
-                    lower_expr(ctx, &args[1])?
-                } else {
-                    "302.0".to_string()
-                };
-                let handle = ctx.block().call(
-                    DOUBLE,
-                    "js_response_static_redirect",
-                    &[(I64, &url_ptr), (DOUBLE, &status)],
-                );
-                return Ok(Some(handle));
-            }
-            _ => {}
-        }
-    }
-
-    // ── axios: static method calls (axios.get/post/put/delete/patch) ──
-    // Must be before the receiver guard — these are receiver-less calls.
-    if module == "axios" && object.is_none() {
-        let url_box = if !args.is_empty() {
-            lower_expr(ctx, &args[0])?
-        } else {
-            double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-        };
-        let blk = ctx.block();
-        let url_handle = unbox_to_i64(blk, &url_box);
-        match method {
-            "get" => {
-                let promise = blk.call(I64, "js_axios_get", &[(I64, &url_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "delete" => {
-                let promise = blk.call(I64, "js_axios_delete", &[(I64, &url_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "post" | "put" | "patch" => {
-                // #598: pass the body as a NaN-boxed f64 instead of
-                // unboxing to i64. Pre-fix the unbox produced a raw
-                // pointer the runtime read as `*const StringHeader`
-                // — for an object literal the pointer was a real
-                // ObjectHeader, the runtime read its bytes as a
-                // StringHeader (length / refcount / data prefix),
-                // and the request body became `^@^B^@^@H...` (the
-                // ObjectHeader struct followed by the first character
-                // of the stringified field). The runtime side now
-                // detects strings vs everything-else via the NaN-box
-                // tag and routes through `js_json_stringify`.
-                let body_box = if args.len() > 1 {
-                    lower_expr(ctx, &args[1])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                let rt_fn = match method {
-                    "post" => "js_axios_post",
-                    "put" => "js_axios_put",
-                    _ => "js_axios_patch",
-                };
-                let promise =
-                    ctx.block()
-                        .call(I64, rt_fn, &[(I64, &url_handle), (DOUBLE, &body_box)]);
-                return Ok(Some(nanbox_pointer_inline(ctx.block(), &promise)));
-            }
-            _ => {}
-        }
-    }
-
-    // Everything below needs a receiver.
-    let Some(recv) = object else {
-        return Ok(None);
-    };
-
-    // ── Headers method dispatch ──
-    if module == "Headers" {
-        let h_handle = lower_expr(ctx, recv)?;
-        match method {
-            "set" | "append" => {
-                if args.len() < 2 {
-                    return Ok(Some(double_literal(0.0)));
-                }
-                let key_ptr = get_raw_string_ptr(ctx, &args[0])?;
-                let val_ptr = get_raw_string_ptr(ctx, &args[1])?;
-                ctx.block().call(
-                    DOUBLE,
-                    "js_headers_set",
-                    &[(DOUBLE, &h_handle), (I64, &key_ptr), (I64, &val_ptr)],
-                );
-                return Ok(Some(double_literal(f64::from_bits(
-                    crate::nanbox::TAG_UNDEFINED,
-                ))));
-            }
-            "get" => {
-                if args.is_empty() {
-                    return Ok(Some(double_literal(0.0)));
-                }
-                let key_ptr = get_raw_string_ptr(ctx, &args[0])?;
-                let str_ptr = ctx.block().call(
-                    I64,
-                    "js_headers_get",
-                    &[(DOUBLE, &h_handle), (I64, &key_ptr)],
-                );
-                let blk = ctx.block();
-                return Ok(Some(nanbox_string_inline(blk, &str_ptr)));
-            }
-            "has" => {
-                if args.is_empty() {
-                    return Ok(Some(double_literal(f64::from_bits(
-                        crate::nanbox::TAG_FALSE,
-                    ))));
-                }
-                let key_ptr = get_raw_string_ptr(ctx, &args[0])?;
-                let out = ctx.block().call(
-                    DOUBLE,
-                    "js_headers_has",
-                    &[(DOUBLE, &h_handle), (I64, &key_ptr)],
-                );
-                return Ok(Some(out));
-            }
-            "delete" => {
-                if args.is_empty() {
-                    return Ok(Some(double_literal(f64::from_bits(
-                        crate::nanbox::TAG_UNDEFINED,
-                    ))));
-                }
-                let key_ptr = get_raw_string_ptr(ctx, &args[0])?;
-                ctx.block().call(
-                    DOUBLE,
-                    "js_headers_delete",
-                    &[(DOUBLE, &h_handle), (I64, &key_ptr)],
-                );
-                return Ok(Some(double_literal(f64::from_bits(
-                    crate::nanbox::TAG_UNDEFINED,
-                ))));
-            }
-            "forEach" => {
-                if args.is_empty() {
-                    return Ok(Some(double_literal(0.0)));
-                }
-                let cb = lower_expr(ctx, &args[0])?;
-                ctx.block().call(
-                    DOUBLE,
-                    "js_headers_for_each",
-                    &[(DOUBLE, &h_handle), (DOUBLE, &cb)],
-                );
-                return Ok(Some(double_literal(f64::from_bits(
-                    crate::nanbox::TAG_UNDEFINED,
-                ))));
-            }
-            // `headers.keys()` / `.values()` / `.entries()` return arrays
-            // sorted by header name (WHATWG Fetch spec). The arrays are
-            // themselves iterable via the array Symbol.iterator, so
-            // `for…of`, spread, and `Array.from` all work for free
-            // (refs #576).
-            "keys" => {
-                let arr = ctx
-                    .block()
-                    .call(DOUBLE, "js_headers_keys", &[(DOUBLE, &h_handle)]);
-                return Ok(Some(arr));
-            }
-            "values" => {
-                let arr = ctx
-                    .block()
-                    .call(DOUBLE, "js_headers_values", &[(DOUBLE, &h_handle)]);
-                return Ok(Some(arr));
-            }
-            "entries" => {
-                let arr = ctx
-                    .block()
-                    .call(DOUBLE, "js_headers_entries", &[(DOUBLE, &h_handle)]);
-                return Ok(Some(arr));
-            }
-            _ => return Ok(None),
-        }
-    }
-
-    // ── Request property getters ──
-    if module == "Request" {
-        let h_handle = lower_expr(ctx, recv)?;
-        match method {
-            "url" => {
-                let str_ptr = ctx
-                    .block()
-                    .call(I64, "js_request_get_url", &[(DOUBLE, &h_handle)]);
-                let blk = ctx.block();
-                return Ok(Some(nanbox_string_inline(blk, &str_ptr)));
-            }
-            "method" => {
-                let str_ptr =
-                    ctx.block()
-                        .call(I64, "js_request_get_method", &[(DOUBLE, &h_handle)]);
-                let blk = ctx.block();
-                return Ok(Some(nanbox_string_inline(blk, &str_ptr)));
-            }
-            "body" => {
-                let val = ctx
-                    .block()
-                    .call(DOUBLE, "js_request_get_body", &[(DOUBLE, &h_handle)]);
-                return Ok(Some(val));
-            }
-            _ => return Ok(None),
-        }
-    }
-
-    // ── Response methods / property getters ──
-    if module == "fetch" {
-        // Lower the receiver once. It's a NaN-boxed POINTER_TAG handle (Phase 1
-        // of the handle-NaN-boxing unification, refs #421) — accessors unbox
-        // via `handle_id` on entry, so codegen passes recv_handle through as
-        // DOUBLE without any fptosi/bitcast conversion. May also be a chained
-        // result from `.headers` / `.clone()` — those cases are recognised at
-        // the Call callsite in lower_call.
-        let recv_handle = lower_expr(ctx, recv)?;
-        match method {
-            "text" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_fetch_response_text", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "json" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_fetch_response_json", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "status" => {
-                let blk = ctx.block();
-                let status = blk.call(
-                    DOUBLE,
-                    "js_fetch_response_status",
-                    &[(DOUBLE, &recv_handle)],
-                );
-                return Ok(Some(status));
-            }
-            "statusText" => {
-                let blk = ctx.block();
-                let str_ptr = blk.call(
-                    I64,
-                    "js_fetch_response_status_text",
-                    &[(DOUBLE, &recv_handle)],
-                );
-                return Ok(Some(nanbox_string_inline(blk, &str_ptr)));
-            }
-            "ok" => {
-                // js_fetch_response_ok returns 1.0 or 0.0 as f64. Map to
-                // TAG_TRUE/TAG_FALSE so console.log prints "true"/"false".
-                let blk = ctx.block();
-                let raw = blk.call(DOUBLE, "js_fetch_response_ok", &[(DOUBLE, &recv_handle)]);
-                let cmp = blk.fcmp("une", &raw, "0.0");
-                let tagged = blk.select(
-                    crate::types::I1,
-                    &cmp,
-                    I64,
-                    crate::nanbox::TAG_TRUE_I64,
-                    crate::nanbox::TAG_FALSE_I64,
-                );
-                return Ok(Some(blk.bitcast_i64_to_double(&tagged)));
-            }
-            "headers" => {
-                let out =
-                    ctx.block()
-                        .call(DOUBLE, "js_response_get_headers", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(out));
-            }
-            "clone" => {
-                let out = ctx
-                    .block()
-                    .call(DOUBLE, "js_response_clone", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(out));
-            }
-            "arrayBuffer" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_response_array_buffer", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "blob" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_response_blob", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            // Issue #237: response.body — returns ReadableStream over the
-            // buffered body bytes. Property access lowers as a zero-arg
-            // method call here, same as response.headers above.
-            "body" => {
-                let h = ctx
-                    .block()
-                    .call(DOUBLE, "js_response_body", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(h));
-            }
-            _ => return Ok(None),
-        }
-    }
-
-    // ── Blob instance methods + property getters (issue #234) ──
-    // The receiver is a numeric Blob handle (registry id) carried as f64,
-    // mirroring the Response handle ABI. Locals are tagged blob::Blob via
-    // `register_native_instance` in `destructuring.rs`.
-    if module == "blob" {
-        let recv_handle = lower_expr(ctx, recv)?;
-        match method {
-            "size" => {
-                let blk = ctx.block();
-                let n = blk.call(DOUBLE, "js_blob_size", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(n));
-            }
-            "type" => {
-                let str_ptr = ctx
-                    .block()
-                    .call(I64, "js_blob_type", &[(DOUBLE, &recv_handle)]);
-                let blk = ctx.block();
-                return Ok(Some(nanbox_string_inline(blk, &str_ptr)));
-            }
-            "arrayBuffer" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_blob_array_buffer", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "bytes" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_blob_bytes", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "text" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_blob_text", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "slice" => {
-                // slice(start?, end?, type?) — missing numeric args use
-                // canonical f64::NAN as sentinel; missing type uses null
-                // pointer (0). Runtime `js_blob_slice` checks `is_nan()`
-                // / `type_ptr.is_null()` to apply WHATWG defaults
-                // (start=0, end=len, type="").
-                let start = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::NAN)
-                };
-                let end = if args.len() >= 2 {
-                    lower_expr(ctx, &args[1])?
-                } else {
-                    double_literal(f64::NAN)
-                };
-                let type_ptr = if args.len() >= 3 {
-                    get_raw_string_ptr(ctx, &args[2])?
-                } else {
-                    "0".to_string()
-                };
-                let new_handle = ctx.block().call(
-                    DOUBLE,
-                    "js_blob_slice",
-                    &[
-                        (DOUBLE, &recv_handle),
-                        (DOUBLE, &start),
-                        (DOUBLE, &end),
-                        (I64, &type_ptr),
-                    ],
-                );
-                return Ok(Some(new_handle));
-            }
-            // Issue #237: blob.stream() — returns ReadableStream over the
-            // blob's bytes. Single-chunk; closes after one read.
-            "stream" => {
-                let h = ctx
-                    .block()
-                    .call(DOUBLE, "js_blob_stream", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(h));
-            }
-            _ => return Ok(None),
-        }
-    }
-
-    // ─────────────────────────────────────────────────────────────────
-    // Web Streams API (issue #237)
-    // The receivers are numeric registry-id handles carried as f64,
-    // mirroring the Blob/Response handle ABI. Locals are tagged
-    // (module, class_name) by `register_native_instance` in
-    // `destructuring.rs`.
-    // ─────────────────────────────────────────────────────────────────
-
-    if module == "readable_stream" {
-        let recv_handle_raw = lower_expr(ctx, recv)?;
-        // Issue #562: subclass instances stash the handle id under
-        // `__perry_stream_handle__`; bare numeric handles pass through
-        // unchanged. Cheap (one runtime call) and applied uniformly so
-        // the FFIs below see a clean registry id either way.
-        let recv_handle = ctx.block().call(
-            DOUBLE,
-            "js_stream_unwrap_handle",
-            &[(DOUBLE, &recv_handle_raw)],
-        );
-        match method {
-            "getReader" => {
-                let h = ctx.block().call(
-                    DOUBLE,
-                    "js_readable_stream_get_reader",
-                    &[(DOUBLE, &recv_handle)],
-                );
-                return Ok(Some(h));
-            }
-            "cancel" => {
-                let reason = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                let blk = ctx.block();
-                let promise = blk.call(
-                    I64,
-                    "js_readable_stream_cancel",
-                    &[(DOUBLE, &recv_handle), (DOUBLE, &reason)],
-                );
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "tee" => {
-                let h =
-                    ctx.block()
-                        .call(DOUBLE, "js_readable_stream_tee", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(h));
-            }
-            "pipeTo" => {
-                let dest_raw = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                // Issue #562: `dest` may be a subclass instance — unwrap.
-                let dest =
-                    ctx.block()
-                        .call(DOUBLE, "js_stream_unwrap_handle", &[(DOUBLE, &dest_raw)]);
-                let blk = ctx.block();
-                let promise = blk.call(
-                    I64,
-                    "js_readable_stream_pipe_to",
-                    &[(DOUBLE, &recv_handle), (DOUBLE, &dest)],
-                );
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "pipeThrough" => {
-                // pipeThrough(transform) — transform has .readable / .writable.
-                // We need both sub-handles. Lower the transform once, then
-                // call js_transform_stream_writable / _readable to extract.
-                let transform_raw = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                // Issue #562: `transform` may be a subclass instance — unwrap.
-                let transform = ctx.block().call(
-                    DOUBLE,
-                    "js_stream_unwrap_handle",
-                    &[(DOUBLE, &transform_raw)],
-                );
-                let writable = ctx.block().call(
-                    DOUBLE,
-                    "js_transform_stream_writable",
-                    &[(DOUBLE, &transform)],
-                );
-                let readable = ctx.block().call(
-                    DOUBLE,
-                    "js_transform_stream_readable",
-                    &[(DOUBLE, &transform)],
-                );
-                let new_h = ctx.block().call(
-                    DOUBLE,
-                    "js_readable_stream_pipe_through",
-                    &[
-                        (DOUBLE, &recv_handle),
-                        (DOUBLE, &writable),
-                        (DOUBLE, &readable),
-                    ],
-                );
-                return Ok(Some(new_h));
-            }
-            "locked" => {
-                let v = ctx.block().call(
-                    DOUBLE,
-                    "js_readable_stream_locked",
-                    &[(DOUBLE, &recv_handle)],
-                );
-                return Ok(Some(v));
-            }
-            // ReadableStreamDefaultController on the same handle:
-            "enqueue" => {
-                let chunk = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                let v = ctx.block().call(
-                    DOUBLE,
-                    "js_readable_stream_controller_enqueue",
-                    &[(DOUBLE, &recv_handle), (DOUBLE, &chunk)],
-                );
-                return Ok(Some(v));
-            }
-            "close" => {
-                let v = ctx.block().call(
-                    DOUBLE,
-                    "js_readable_stream_controller_close",
-                    &[(DOUBLE, &recv_handle)],
-                );
-                return Ok(Some(v));
-            }
-            "error" => {
-                let reason = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                let v = ctx.block().call(
-                    DOUBLE,
-                    "js_readable_stream_controller_error",
-                    &[(DOUBLE, &recv_handle), (DOUBLE, &reason)],
-                );
-                return Ok(Some(v));
-            }
-            "desiredSize" => {
-                let v = ctx.block().call(
-                    DOUBLE,
-                    "js_readable_stream_controller_desired_size",
-                    &[(DOUBLE, &recv_handle)],
-                );
-                return Ok(Some(v));
-            }
-            _ => return Ok(None),
-        }
-    }
-
-    if module == "readable_stream_reader" {
-        let recv_handle = lower_expr(ctx, recv)?;
-        match method {
-            "read" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_reader_read", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "releaseLock" => {
-                let v =
-                    ctx.block()
-                        .call(DOUBLE, "js_reader_release_lock", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(v));
-            }
-            "cancel" => {
-                let reason = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                let blk = ctx.block();
-                let promise = blk.call(
-                    I64,
-                    "js_reader_cancel",
-                    &[(DOUBLE, &recv_handle), (DOUBLE, &reason)],
-                );
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "closed" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_reader_closed", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            _ => return Ok(None),
-        }
-    }
-
-    if module == "writable_stream" {
-        let recv_handle_raw = lower_expr(ctx, recv)?;
-        // Issue #562: subclass instances unwrap to a numeric handle.
-        let recv_handle = ctx.block().call(
-            DOUBLE,
-            "js_stream_unwrap_handle",
-            &[(DOUBLE, &recv_handle_raw)],
-        );
-        match method {
-            "getWriter" => {
-                let h = ctx.block().call(
-                    DOUBLE,
-                    "js_writable_stream_get_writer",
-                    &[(DOUBLE, &recv_handle)],
-                );
-                return Ok(Some(h));
-            }
-            "abort" => {
-                let reason = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                let blk = ctx.block();
-                let promise = blk.call(
-                    I64,
-                    "js_writable_stream_abort",
-                    &[(DOUBLE, &recv_handle), (DOUBLE, &reason)],
-                );
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "close" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_writable_stream_close", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "locked" => {
-                let v = ctx.block().call(
-                    DOUBLE,
-                    "js_writable_stream_locked",
-                    &[(DOUBLE, &recv_handle)],
-                );
-                return Ok(Some(v));
-            }
-            _ => return Ok(None),
-        }
-    }
-
-    if module == "writable_stream_writer" {
-        let recv_handle = lower_expr(ctx, recv)?;
-        match method {
-            "write" => {
-                let chunk = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                let blk = ctx.block();
-                let promise = blk.call(
-                    I64,
-                    "js_writer_write",
-                    &[(DOUBLE, &recv_handle), (DOUBLE, &chunk)],
-                );
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "close" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_writer_close", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "abort" => {
-                let reason = if !args.is_empty() {
-                    lower_expr(ctx, &args[0])?
-                } else {
-                    double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED))
-                };
-                let blk = ctx.block();
-                let promise = blk.call(
-                    I64,
-                    "js_writer_abort",
-                    &[(DOUBLE, &recv_handle), (DOUBLE, &reason)],
-                );
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "releaseLock" => {
-                let v =
-                    ctx.block()
-                        .call(DOUBLE, "js_writer_release_lock", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(v));
-            }
-            "closed" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_writer_closed", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "ready" => {
-                let blk = ctx.block();
-                let promise = blk.call(I64, "js_writer_ready", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(nanbox_pointer_inline(blk, &promise)));
-            }
-            "desiredSize" => {
-                let v =
-                    ctx.block()
-                        .call(DOUBLE, "js_writer_desired_size", &[(DOUBLE, &recv_handle)]);
-                return Ok(Some(v));
-            }
-            _ => return Ok(None),
-        }
-    }
-
-    if module == "transform_stream" {
-        let recv_handle_raw = lower_expr(ctx, recv)?;
-        // Issue #562: subclass instances unwrap to a numeric handle.
-        let recv_handle = ctx.block().call(
-            DOUBLE,
-            "js_stream_unwrap_handle",
-            &[(DOUBLE, &recv_handle_raw)],
-        );
-        match method {
-            "readable" => {
-                let v = ctx.block().call(
-                    DOUBLE,
-                    "js_transform_stream_readable",
-                    &[(DOUBLE, &recv_handle)],
-                );
-                return Ok(Some(v));
-            }
-            "writable" => {
-                let v = ctx.block().call(
-                    DOUBLE,
-                    "js_transform_stream_writable",
-                    &[(DOUBLE, &recv_handle)],
-                );
-                return Ok(Some(v));
-            }
-            _ => return Ok(None),
-        }
-    }
-
-    // ── axios: response property access (response.status, .data, .statusText, .headers) ──
-    if module == "axios" {
-        if let Some(recv) = object {
-            let recv_handle = lower_expr(ctx, recv)?;
-            let blk = ctx.block();
-            // The awaited axios response is a Handle (i64) NaN-boxed via
-            // `JsValue::from_object_ptr(handle as *mut ())` (POINTER_TAG |
-            // (handle & POINTER_MASK)). Use `unbox_to_i64` to strip the
-            // tag and recover the bare handle id; calling
-            // `bitcast_double_to_i64` alone leaves the upper-16 tag bits
-            // and the runtime's `get_handle::<AxiosResponseHandle>` lookup
-            // misses, returning 0 / undefined for every property. (#604
-            // followup — only surfaced once the listen() hang was fixed.)
-            let h_i64 = unbox_to_i64(blk, &recv_handle);
-            match method {
-                "status" => {
-                    let status = blk.call(DOUBLE, "js_axios_response_status", &[(I64, &h_i64)]);
-                    return Ok(Some(status));
-                }
-                "statusText" => {
-                    let str_ptr = blk.call(I64, "js_axios_response_status_text", &[(I64, &h_i64)]);
-                    return Ok(Some(nanbox_string_inline(blk, &str_ptr)));
-                }
-                "data" => {
-                    // Use the auto-parsed variant (JSON when the body
-                    // looks like JSON, raw string otherwise) so
-                    // `r.data.ok` / `r.data[0]` work the same way as
-                    // in npm `axios`. The function returns a NaN-boxed
-                    // f64 directly; no need to nanbox here. (#604
-                    // followup — only surfaced once listen() hang fix
-                    // unblocked the axios chain.)
-                    let v = blk.call(DOUBLE, "js_axios_response_data_parsed", &[(I64, &h_i64)]);
-                    return Ok(Some(v));
-                }
-                _ => {}
-            }
-        }
-    }
-
-    Ok(None)
-}
 
 /// Static dispatch table for perry/ui receiver-less calls. Covers the
 /// constructors + setters mango uses, plus the most common widgets from
